Do not lose wakeups for block device workers (#973)
The QueueMinder's `notify_workers` is the sole indicator that a worker
thread is asleep or not, so `take_notifications` transfers the
responsibility to handle these bits to the caller, which
`flush_notifications` did not fully handle.

Return remaining bits indicating idle workers to QueueMinder, so when
there's enough work to justify waking them, we actually do.

In the process, fix a less exciting bug where attempting to spuriously
wake already-woken worker threads would result in waking fewer workers
than desired.

And, along the way, add a few new probes that were useful in debugging.